Import vocabulary from csv file (YTI-303)
Read csv file and convert it into JSON. Parse concepts from JSON and show them in the preview modal. If user accepts the import then convert concepts from JSON into concept nodes and save these nodes.

The preview modal is not visually finished yet so this is a demo version.