Impersonate user
Desktop: Show a list of fakeable users in a popup menu that appears
when hovering or focusing login button (or username/logout when logged
in).

Mobile: Show a list of fakeabe users at the bottom of hamburger menu.

This works only if fakeable users are found from back-end. That means
only development environments.