Fix fonts
Source Sans Pro is loaded from Google Fonts API via CSS @import
not SASS @import. Thus is is not transpiled by SASS. And finally
the browser should load that CSS file.

In CSS, @imports must be at the top of the file but somehow SASS
didn't put it first. And then browser ignored it silently.

This fix does some migrations (as @import is deprecated in SASS now)
and somehow that CSS @import is now at the top of the file. I don't
know why this works. I need to test this in dev server before I dare
to say if this works or not.

YTI-1662