Fix some i18n issues from page title
In some cases, page title was not translated correctly.

For example, in concept page the title should contain prefLabel
from prefLabelXl term. Terms are language spesific so one term
contains prefLabel in one language. Thus is is not good enough to
get prefLabel from the first prefLabelXl but instead from all of
them.

In addition, refactored rendering of page title into single
component. Testing it was very hard (because of how next/head
works) so no tests for it in this commit.