Use environment variables with Matomo (#139)
Environment variables can be accessed only on the server. If they are
needed in the browser they must be passed via props. Made this easier
by creating CommonContextProvider that is used on every page and
takes all the page props. Currently, other components can access these
values by simply using this context.

The first two use cases for CommonContextProvider are Matomo and
guessing results of media queries in SSR. Later it would be useful
with e.g. feature flags.

YTI-1800