Use styled-components' transient props
When passing props to styled components they are normally passed
through to the DOM. This caused warnings in tests as custom DOM
props must use lower case and our props were using camel case.

Fixed this by renaming all the props with $. Now they are not
passed down and no more these warnings in the tests.