pass JSESSIONID to API calls in SSR
This should allow us to use the same authenticated API endpoints in both
SSR and browser.

extraArgument is used to pass NextIronContext to baseQuery.
axiosBaseQuery can be provided with a prepareHeaders function, which has
access to the context (and thus the session), and allows for modifying
the request headers.

Essentially we can now read JSESSIONID previously stored in the session,
and pass it along to the axios request as a cookie.

Reference implementation without all the project specific stuff exists at
https://github.com/urhot/nextjs-rtk-query-ssr