Fix pluralizations from translations (#180)
Convert translations into v4 format. It seems that we are actually using the v4
format instead of the v3 format. I don't know why/how I thought we should use
the v3 format earlier. :man_shrugging: 

In addition, fixed some places where pluralization was not used as the
parameter was named incorrectly. It must be `count`, not `amount` or
`number`.

YTI-2007