Confirm before leaving page with unsaved form (#190)
Create useConfirmBeforeLeavingPage hook that registers beforeUnload
and routeChangeStart event listeners. When either happens it shows
a confirmation dialog. The useConfirmBeforeLeavingPage hook also
returns enable and disable functions which can be used to enable and
disable this behavior at will.

Also injected this hook in the create and edit forms of terminology,
concept, and collection.

YTI-1996, YTI-1940, YTI-2005