Fix don't confirm before leaving page without form
When confirm is correctly enabled and user navigates to another
page using next/router the confirmation is triggered. After this
the confirmation is not disabled and so it was still enabled even
if the user has navigated to a page without form at all.

Fixed this by enabling disabling confirmation automatically when
the page component is is unmounted.