fix next.js error related to SSR
In production build, we would frequently get errors:

* Minified React error #425;
* Minified React error #418;
* Minified React error #423;

Which translate to:

* Text content does not match server-rendered HTML
* Hydration failed because the initial UI does not match what was
  rendered on the server.
* There was an error while hydrating. Because the error happened outside
  of a Suspense boundary, the entire root will switch to client
  rendering.

Some possible reasons for these errors are found in nextjs
documentation:

https://nextjs.org/docs/messages/react-hydration-error

In this case, the culprit appears to be the FormattedDate component,
which uses dayjs to for date formatting.

As a workaround, the formatting is now done only on client-side.