fix redirection after model creation
previously, if the model page had a "new" query parameter, router would
replace the url to not include it.

This breaks things, since new models are drafts, but urls without "new"
or "draft" are considered non-drafts.

To fix this, the url replacement still remains, but now changes the
"new" parameter to "draft".