fix for resource being saved to a wrong datamodel
Browsing to an attribute/association would store the modelId in
the "active" state, which would then be be considered the current
modelId when creating a new attribute/association.

This means saving the new attribute/association was being done to
whatever modelId was previously set in state.

This fix will reset the active state when browsing to the
attribute/association create form, and the resource should now be saved
to the proper datamodel.